Fixup errors with unit test mock repo
